% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrplot.R
\name{lrplot}
\alias{lrplot}
\title{Creates a point plot with a linear regression between weight and hindfoot length}
\usage{
lrplot(surveys)
}
\arguments{
\item{a}{The dataset used}
}
\value{
plot The resulting plot with a linear regression between weight and hindfoot length
}
\description{
Creates a point plot with a linear regression between weight and hindfoot length
}
