% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Reads the given file in Tidyverse wihout NA values}
\usage{
read("cloud/project/data/surveys.csv")
read("cloud/project/data/Butterfly_data.csv")
}
\arguments{
\item{x}{File being read}
}
\value{
fileread The file is read
}
\description{
Reads the given file in Tidyverse wihout NA values
}
