% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Creates a linear regression between weight and hindfoot length of a dataset}
\usage{
linreg(surveys)
}
\arguments{
\item{z}{The dataset being used}
}
\value{
lr The linear regression between the two values
}
\description{
Creates a linear regression between weight and hindfoot length of a dataset
}
