% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplot.R
\name{pplot}
\alias{pplot}
\title{Creates a point plot between weight and hindfoot length of a given dataset}
\usage{
pplot(surveys)
}
\arguments{
\item{a}{Dataset being used}
}
\value{
plot The resulting plot showing weight against hindfoot length of the dataset
}
\description{
Creates a point plot between weight and hindfoot length of a given dataset
}
